refactor: 스토리북 타이틀 정리 및 폴더 이동
