+++ src/containers/CardNumberInput/CardNumber.jsx
@@ -3,10 +3,10 @@ import ErrorMessage from 'containers/ErrorMessage/ErrorMessage';
 import validator from 'lib/validations';
 import { CardDispatchContext, CardStateContext } from 'store/card/CardContext';
 import { TYPES } from 'store/card/types';
-import AutoFocusInputBox from 'common/AutoFocusInputBox';
+import AutoFocusInputBox from 'common/AutoFocus/AutoFocusInputBox';
 import Label from 'components/Label/Label';
 import styled from 'styled-components';
-import InputBox from 'components/InputBox/InputBox';
+import InputBox from 'common/InputBox/InputBox';
 import { InputBasic } from 'components/Input/Input';
 
 export default function CardNumber({ color, setIsListModalOpen }) {
